+++ src/main/java/fr/sma/adventofcode/resolve/processor/lambda/Instruction.java
@@ -1,5 +1,5 @@
 package fr.sma.adventofcode.resolve.processor.lambda;
 
 public interface Instruction {
-	void execute(int[] register);
+	void execute(long[] register);
 }
